Simultaneously display arrow and rudimentary random color assignment/glitchy blink patterns.
The essential change in this patch is the use of ranges in the LED strip collection
instead of slices of the same. The previous approach resulted in multiple mutable references
to the same LED object (i.e., strip collection item) or immutable uses (in ws.write) with
mutable references still in scope, blocking compilation. More than anything, the slices
served to improve readability of the code (e.g., arrow_leds); to hold on to some of that
legibility, constants have been created for ranges in the strip (e.g., strip[ARROW_LED_RANGE]).